ARPG + Extra

+++ DOTA/DOTA/DOTA/Base.cpp
@@ -5,8 +5,8 @@ void Base::InitialiseBase() {
 }
 
 void Base::checkBarracks(std::array<MeleeBarracks, 3> &m, std::array<RangedBarracks, 3> &r) {
-	short meleeDestroyed = 0;
-	short rangeDestroyed = 0;
+	int meleeDestroyed = 0;
+	int rangeDestroyed = 0;
 	for(unsigned int i = 0; i < 3; i++) {
 		if(m[i].health == 0)
 			meleeDestroyed++;
