+++ DOTA/DOTA/DOTA/LaneCreep.cpp
@@ -4,7 +4,7 @@ void LaneCreep::CheckBase(Base &b) {
 	return;
 }
 
-short LaneCreep::AutoAttack() {
+int LaneCreep::AutoAttack() {
 	std::cout << "Lane creep hit!\n";
 	return m_Damage;
 }
