+++ DOTA/DOTA/DOTA/Player.h
@@ -8,7 +8,7 @@
 struct Player {
 	int teamID = 0, creepsKilled = 0;
 	char choice = ' ';
-	short chanceToAttack = 0;
+	int chanceToAttack = 0;
 	bool fighting = false;
 	std::string currentTeam = " ";
 	Hero *currentHero; // Initialise! Axe = default
