+++ DOTA/DOTA/DOTA/Tower.h
@@ -5,7 +5,7 @@
 
 class Tower {
 private:
-	const short attackRange = 2000;
+	const int attackRange = 2000;
 	short damage = 10, health = 2500;
 public:
 	const std::string m_Type = " "; // Dictates what level of tower it is, front, mid, base, ancient
